+++ scripts/reinstall_django17.sh
@@ -1,3 +1,3 @@
 reset
-./uninstall.sh
-./install_django17.sh
+./scripts/uninstall.sh
+./scripts/install_django17.sh