+++ scripts/test3.sh
@@ -1,6 +1,6 @@
 reset
-./uninstall.sh
-./install3.sh
+./scripts/uninstall.sh
+./scripts/install3.sh
 #cd ..
 python examples/simple/manage.py test fobi --traceback
 #cd scripts