+++ scripts/test_pypi.sh
@@ -1,6 +1,6 @@
 reset
-./uninstall.sh
-./install_pypi.sh
+./scripts/uninstall.sh
+./scripts/install_pypi.sh
 #cd ..
 python examples/simple/manage.py test fobi
 #cd scripts