+++ src/fobi/urls/view.py
@@ -4,18 +4,24 @@ __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('urlpatterns',)
 
-from django.conf.urls import patterns, url #include
+from django.conf.urls import url #patterns, include
+from fobi.views import (
+    form_entry_submitted, view_form_entry, form_entry_submitted
+)
 
-
-urlpatterns = patterns('fobi.views',
+#urlpatterns = patterns('fobi.views',
+urlpatterns = [
     # Form submitted success page
-    url(r'^view/submitted/$', view='form_entry_submitted', name='fobi.form_entry_submitted'),
+    url(r'^view/submitted/$', view=form_entry_submitted,
+        name='fobi.form_entry_submitted'),
 
     # View form entry
-    url(r'^view/(?P<form_entry_slug>[\w_\-]+)/$', 'view_form_entry', name='fobi.view_form_entry'),
+    url(r'^view/(?P<form_entry_slug>[\w_\-]+)/$', view_form_entry,
+        name='fobi.view_form_entry'),
 
     # Form submitted success page
-    url(r'^view/(?P<form_entry_slug>[\w_\-]+)/submitted/$', view='form_entry_submitted', name='fobi.form_entry_submitted'),
+    url(r'^view/(?P<form_entry_slug>[\w_\-]+)/submitted/$',
+        view=form_entry_submitted, name='fobi.form_entry_submitted'),
 
     # Create form entry
     #url(r'^forms/create/$', view='create_form_entry', name='fobi.create_form_entry'),
@@ -52,4 +58,5 @@ urlpatterns = patterns('fobi.views',
 
     # Dashboard
     #url(r'^$', view='dashboard', name='fobi.dashboard'),
-)
+#)
+]