+++ scripts/install_django17.sh
@@ -1,6 +1,6 @@
 #pip install -r examples/requirements.txt --allow-all-external --allow-unverified django-admin-tools
 #cd ..
-pip install -r examples/requirements.txt
+pip install -r examples/requirements_django17.txt
 python setup.py install
 mkdir -p examples/logs examples/db examples/media examples/media/static examples/media/fobi_plugins/content_image
 mkdir -p examples/media/fobi_plugins/file